Refactor: comment util 폴더 수정
