+++ src/modules/CommentList/Comment/Comment.client.tsx
@@ -6,7 +6,7 @@ import {
   useDeleteComment,
   useDeleteCommentLike,
   usePostLikeComment,
-} from '~/api/domain/comment.api';
+} from '~/api/domain/comment/comment.api';
 import {
   Content,
   DeleteButton,
